pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "influencerpy"
version = "0.1.0"
description = "Premium Social Media Automation CLI"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
  { name = "Cristian Leo", email = "cristianleo120@gmail.com" },
]
dependencies = [
  "typer",
  "tweepy",
  "python-telegram-bot",
  "python-dotenv",
  "rich",
  "sqlmodel",
  "requests",
  "google-generativeai",
  "questionary",
  "pyfiglet",
  "mkdocs-material",
  "arxiv",
  "pytz",
  "beautifulsoup4",
  "apscheduler",
  "strands-agents-tools",
  "strands-agents[anthropic]",
  "strands-agents[gemini]",
  "strands-agents[otel]",
  "strands-agents-tools[rss]",
  "strands-agents-tools[local_chromium_browser]",
  "langfuse",
  "sentence-transformers",
  "torch>=2.0.0,<2.3.0",
  "numpy<2",
]

[project.scripts]
influencerpy = "influencerpy.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/influencerpy"]
